refactor(models): extract agents/available/aliases tabs, thin the shell (phase 2b)
Final phase of the models-page.tsx split. Extracts the last three tab
branches and the shared render helpers, leaving a lean page shell.

- agents-tab.tsx (AgentsTab) — global defaults card, fallback-models
  reorderable editor, per-agent config table.
- available-models-tab.tsx (AvailableModelsTab) — refresh/cache header,
  loading/error states, provider-grouped model cards; owns TIER_STYLES +
  formatRelativeTime (their only consumer).
- aliases-tab.tsx (AliasesTab) — alias table + add-alias form.
- models-page-shared.tsx — TableSkeleton + InlineEmpty, shared by the agents
  and aliases tabs.

models-page.tsx is now a 76-line shell: PluginHeader, error/restart banners,
the tab bar, and the five `<XTab m={m} />` renders, all fed by useModelsData().

Pure behavior-preserving relocation — every tab's JSX is byte-identical.

This completes the models-page decomposition: 1,205 -> 76-line shell across 8
cohesive files (hook + 5 tab components + shared helpers).

Verified: typecheck + eslint + models-page.test (10/0, renders all five tabs)
+ full suite (5124/0) + binary build + isolated-home boot smoke (Models plugin
loads; rebuilt client.js → 200).
