+++ plugins/models/components/models-page.tsx
@@ -1,50 +1,17 @@
 'use client'
 
-// External
-import { ArrowDown, ArrowUp, Plus, X, Users, RefreshCw, AlertTriangle } from 'lucide-react'
 // Internal
 import { Button } from "@makinbakin/sdk/ui"
 import { PluginHeader } from "@makinbakin/sdk/components"
 import { ErrorBanner } from "@makinbakin/sdk/components"
-import { EmptyState } from "@makinbakin/sdk/components"
 import { UnderlineTabs } from "@makinbakin/sdk/components"
-import { Input } from "@makinbakin/sdk/ui"
-import { Badge } from "@makinbakin/sdk/ui"
-import { Skeleton } from "@makinbakin/sdk/ui"
-import {
-  Table, TableHeader, TableBody, TableHead, TableRow, TableCell,
-} from "@makinbakin/sdk/ui"
-import { AgentAvatar } from "@makinbakin/sdk/components"
-import { ModelSelect } from "@makinbakin/sdk/components"
 // Relative
-import { BrandIcon } from './brand-icon'
 import { useModelsData } from './use-models-data'
+import { AgentsTab } from './agents-tab'
+import { AvailableModelsTab } from './available-models-tab'
+import { AliasesTab } from './aliases-tab'
 import { RoutingTab } from './routing-tab'
 import { SpendTab } from './spend-tab'
-
-// ---------------------------------------------------------------------------
-// Relative-time formatter — under 1h: minutes, under 24h: hours, else date.
-// ---------------------------------------------------------------------------
-function formatRelativeTime(ts: number | null): string {
-  if (!ts) return 'never'
-  const delta = Date.now() - ts
-  const seconds = Math.max(0, Math.floor(delta / 1000))
-  if (seconds < 60) return 'just now'
-  const minutes = Math.floor(seconds / 60)
-  if (minutes < 60) return `${minutes}m ago`
-  const hours = Math.floor(minutes / 60)
-  if (hours < 24) return `${hours}h ago`
-  return new Date(ts).toLocaleDateString(undefined, { month: 'short', day: 'numeric' })
-}
-
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-const TIER_STYLES: Record<string, string> = {
-  budget: 'bg-green-500/10 text-green-400',
-  standard: 'bg-blue-500/10 text-blue-400',
-  premium: 'bg-purple-500/10 text-purple-400',
-}
 
 const TABS = [
   { id: 'agents', label: 'Agent Config' },
@@ -55,57 +22,12 @@ const TABS = [
 ] as const
 
 // ---------------------------------------------------------------------------
-// Loading skeleton
-// ---------------------------------------------------------------------------
-function TableSkeleton({ rows = 4, cols = 4 }: { rows?: number; cols?: number }) {
-  return (
-    <div className="overflow-hidden rounded-xl border border-border">
-      <Table>
-        <TableHeader>
-          <TableRow className="bg-card">
-            {Array.from({ length: cols }).map((_, i) => (
-              <TableHead key={i}><Skeleton className="h-4 w-20" /></TableHead>
-            ))}
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {Array.from({ length: rows }).map((_, r) => (
-            <TableRow key={r}>
-              {Array.from({ length: cols }).map((_, c) => (
-                <TableCell key={c}><Skeleton className="h-4 w-full" /></TableCell>
-              ))}
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </div>
-  )
-}
-
-function InlineEmpty({ message }: { message: string }) {
-  return (
-    <div className="py-6 text-center text-sm text-muted-foreground">{message}</div>
-  )
-}
-
-// ---------------------------------------------------------------------------
-// Main component
+// Main component — page shell: header, banners, tab bar, and the five tabs
+// (each fed the shared useModelsData() object).
 // ---------------------------------------------------------------------------
 export function ModelsPage() {
   const m = useModelsData()
-  const {
-    tab, setTab,
-    agents, loading, error, saving, runtimeStatus, fetchConfig,
-    availableModels, modelOptions, modelsReady, availableProviders,
-    modelsCached, modelsCachedAt, modelsStale, modelsError, modelsLoaded, refreshing, handleRefresh,
-    aliases, newAliasName, setNewAliasName, newAliasTarget, setNewAliasTarget,
-    addAlias, deleteAlias, prepopulateAliases,
-    pendingOwn, setPendingOwn, pendingSub, setPendingSub,
-    setPendingDefaultModel, setPendingDefaultSubagentModel,
-    pendingFallbackModels, setPendingFallbackModels, fallbackModels,
-    saveAgent, saveAll, saveDefaults, setAsDefault, hasPending, defaultsDirty,
-    effectiveDefaultModel, effectiveDefaultSubagentModel, effectiveFallbackModels, fallbackCandidates,
-  } = m
+  const { tab, setTab, error, fetchConfig, runtimeStatus, modelOptions } = m
 
   return (
     <div className="p-6 flex flex-col flex-1 gap-6">
@@ -143,419 +65,12 @@ export function ModelsPage() {
         onValueChange={(id) => setTab(id as typeof tab)}
       />
 
-
       {/* Tab content */}
-      {tab === 'agents' && (
-        <div className="space-y-4">
-            <div className="rounded-xl border border-border bg-card p-4 space-y-4">
-              <div className="flex items-center justify-between gap-4">
-                <div>
-                  <h3 className="text-sm font-semibold">Global Defaults</h3>
-                  <p className="text-sm text-muted-foreground">
-                    Controls <code className="text-xs">agents.defaults.model.primary</code>, <code className="text-xs">fallbacks</code>, and the default subagent model.
-                  </p>
-                </div>
-                {defaultsDirty && (
-                  <Button onClick={() => saveDefaults()} disabled={!!saving || !modelsReady} size="sm">
-                    Save Defaults
-                  </Button>
-                )}
-              </div>
-
-              <div className="grid gap-4 md:grid-cols-2">
-                <div className="space-y-2">
-                  <label className="text-xs font-medium text-muted-foreground">Default Model</label>
-                  <ModelSelect
-                    value={effectiveDefaultModel}
-                    onChange={(v) => setPendingDefaultModel(v)}
-                    models={modelOptions}
-                  />
-                </div>
-                <div className="space-y-2">
-                  <label className="text-xs font-medium text-muted-foreground">Default Subagent Model</label>
-                  <ModelSelect
-                    value={effectiveDefaultSubagentModel}
-                    onChange={(v) => setPendingDefaultSubagentModel(v === '__default__' ? null : v)}
-                    models={modelOptions}
-                    defaultLabel={`Use primary default (${effectiveDefaultModel})`}
-                  />
-                </div>
-              </div>
-
-              <div className="space-y-2">
-                <div className="flex items-center justify-between gap-4">
-                  <label className="text-xs font-medium text-muted-foreground">Fallback Models</label>
-                  <Button
-                    variant="outline"
-                    size="xs"
-                    onClick={() => setPendingFallbackModels([...(pendingFallbackModels ?? fallbackModels), fallbackCandidates[0]?.id ?? ''])}
-                    disabled={fallbackCandidates.length === 0}
-                  >
-                    <Plus className="mr-1 size-3" />
-                    Add Fallback
-                  </Button>
-                </div>
-
-                <div className="space-y-2">
-                  {effectiveFallbackModels.length === 0 ? (
-                    <InlineEmpty message="No fallback models configured." />
-                  ) : (
-                    effectiveFallbackModels.map((modelId, index) => (
-                      <div key={`${modelId}-${index}`} className="flex items-center gap-2">
-                        <span className="w-16 text-xs text-muted-foreground">#{index + 1}</span>
-                        <ModelSelect
-                          value={modelId}
-                          onChange={(value) => {
-                            const next = [...effectiveFallbackModels]
-                            next[index] = value
-                            setPendingFallbackModels([...new Set(next.filter(Boolean).filter((id) => id !== effectiveDefaultModel))])
-                          }}
-                          models={fallbackCandidates}
-                          className="flex-1"
-                        />
-                        <Button
-                          variant="outline"
-                          size="icon-sm"
-                          onClick={() => {
-                            if (index === 0) return
-                            const next = [...effectiveFallbackModels]
-                            ;[next[index - 1], next[index]] = [next[index], next[index - 1]]
-                            setPendingFallbackModels(next)
-                          }}
-                          disabled={index === 0}
-                        >
-                          <ArrowUp className="size-3" />
-                        </Button>
-                        <Button
-                          variant="outline"
-                          size="icon-sm"
-                          onClick={() => {
-                            if (index === effectiveFallbackModels.length - 1) return
-                            const next = [...effectiveFallbackModels]
-                            ;[next[index], next[index + 1]] = [next[index + 1], next[index]]
-                            setPendingFallbackModels(next)
-                          }}
-                          disabled={index === effectiveFallbackModels.length - 1}
-                        >
-                          <ArrowDown className="size-3" />
-                        </Button>
-                        <Button
-                          variant="outline"
-                          size="icon-sm"
-                          onClick={() => setPendingFallbackModels(effectiveFallbackModels.filter((_, i) => i !== index))}
-                        >
-                          <X className="size-3" />
-                        </Button>
-                      </div>
-                    ))
-                  )}
-                </div>
-              </div>
-            </div>
-
-            {hasPending && (
-              <div className="flex justify-end">
-                <Button onClick={saveAll} disabled={!!saving || !modelsReady} size="sm">
-                  Save All
-                </Button>
-              </div>
-            )}
-
-            {loading ? (
-              <TableSkeleton rows={5} cols={4} />
-            ) : agents.length === 0 ? (
-              <EmptyState icon={Users} title="No agents configured in the runtime" />
-            ) : (
-              <div className="overflow-hidden rounded-xl border border-border">
-                <Table>
-                  <TableHeader>
-                    <TableRow className="bg-card">
-                      <TableHead className="w-[140px]">Agent</TableHead>
-                      <TableHead>Own Model</TableHead>
-                      <TableHead>Subagent Model</TableHead>
-                      <TableHead className="w-[100px]" />
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {agents.map((agent) => {
-                      const ownVal = pendingOwn[agent.agentId] ?? (agent.ownModel || '__default__')
-                      const subVal = pendingSub[agent.agentId] ?? (agent.subagentModel || '__default__')
-                      const hasDirty = agent.agentId in pendingOwn || agent.agentId in pendingSub
-                      const isSaving = saving === agent.agentId
-
-                      return (
-                        <TableRow key={agent.agentId}>
-                          <TableCell>
-                            <div className="flex items-center gap-2">
-                              <AgentAvatar agentId={agent.agentId} size="sm" />
-                              <span className="font-medium">{agent.name}</span>
-                            </div>
-                          </TableCell>
-                          <TableCell>
-                            <ModelSelect
-                              value={ownVal}
-                              onChange={(v) => setPendingOwn((p) => ({ ...p, [agent.agentId]: v }))}
-                              models={modelOptions}
-                              defaultLabel={`Default (${agent.defaultModel})`}
-                            />
-                          </TableCell>
-                          <TableCell>
-                            <ModelSelect
-                              value={subVal}
-                              onChange={(v) => setPendingSub((p) => ({ ...p, [agent.agentId]: v }))}
-                              models={modelOptions}
-                              defaultLabel={`Default (${agent.defaultSubagentModel || agent.defaultModel})`}
-                            />
-                          </TableCell>
-                          <TableCell>
-                            {hasDirty && (
-                              <Button
-                                onClick={() => saveAgent(agent.agentId)}
-                                disabled={isSaving || !modelsReady}
-                                size="xs"
-                              >
-                                {isSaving ? 'Saving...' : 'Save'}
-                              </Button>
-                            )}
-                          </TableCell>
-                        </TableRow>
-                      )
-                    })}
-                  </TableBody>
-                </Table>
-              </div>
-            )}
-        </div>
-      )}
-
-      {tab === 'available' && (
-        <div className="space-y-6">
-            {/* Runtime-out-of-sync banner */}
-            {runtimeStatus.restartNeeded && (
-              <div className="flex items-center justify-between gap-3 rounded-md border border-amber-500/40 bg-amber-500/10 px-3 py-2 text-xs text-amber-300">
-                <div className="flex items-center gap-2">
-                  <AlertTriangle className="size-3.5" />
-                  <span>
-                    Runtime config changed since the last runtime restart. Model list may be out of date until you restart.
-                  </span>
-                </div>
-                <Button
-                  size="xs"
-                  variant="outline"
-                  disabled={runtimeStatus.restarting}
-                  onClick={runtimeStatus.restart}
-                >
-                  {runtimeStatus.restarting ? 'Restarting…' : 'Restart runtime'}
-                </Button>
-              </div>
-            )}
-
-            {/* Refresh + cache-age header */}
-            <div className="flex items-center justify-between gap-3">
-              <div className="text-xs text-muted-foreground">
-                {modelsLoaded && modelsCached && modelsCachedAt ? (
-                  <span>
-                    Last refreshed: <span className="font-medium text-foreground">{formatRelativeTime(modelsCachedAt)}</span>
-                    {modelsStale && <span className="ml-2 text-amber-400">(stale — refreshing…)</span>}
-                  </span>
-                ) : modelsLoaded && !modelsCached && availableModels.length > 0 ? (
-                  <span>Just refreshed</span>
-                ) : null}
-              </div>
-              <Button
-                size="sm"
-                variant="outline"
-                onClick={handleRefresh}
-                disabled={refreshing}
-              >
-                <RefreshCw className={`size-3.5 ${refreshing ? 'animate-spin' : ''}`} />
-                <span className="ml-1">{refreshing ? 'Refreshing…' : 'Refresh'}</span>
-              </Button>
-            </div>
-
-            {/* States: loading / error / list */}
-            {!modelsLoaded || (refreshing && availableModels.length === 0) ? (
-              <div className="flex flex-col items-center justify-center rounded-xl border border-border bg-card py-12 gap-3 text-sm text-muted-foreground">
-                <RefreshCw className="size-5 animate-spin text-foreground/60" />
-                <div>Querying runtime adapter — this can take up to 30 seconds on first load.</div>
-              </div>
-            ) : availableModels.length === 0 ? (
-              <div className="rounded-xl border border-border bg-card p-6 space-y-3">
-                <div className="flex items-center gap-2 text-sm text-foreground">
-                  <AlertTriangle className="size-4 text-red-400" />
-                  Could not load models from the runtime.
-                </div>
-                {modelsError && (
-                  <div className="font-mono text-xs text-muted-foreground break-all">{modelsError}</div>
-                )}
-                <Button size="sm" onClick={handleRefresh} disabled={refreshing}>
-                  <RefreshCw className={`size-3.5 ${refreshing ? 'animate-spin' : ''}`} />
-                  <span className="ml-1">Retry</span>
-                </Button>
-              </div>
-            ) : (
-              availableProviders.map((provider) => {
-                const models = modelOptions.filter((m) => m.provider === provider)
-                if (models.length === 0) return null
-                // Provider-level metadata from the first model (they all share provider fields).
-                const providerMeta = models.find((m) => m.providerLabel) ?? models[0]
-                const providerLabel = providerMeta.providerLabel ?? provider.replace(/[-_]/g, ' ')
-                return (
-                  <div key={provider}>
-                    <h3 className="mb-3 text-sm font-semibold uppercase tracking-wider text-muted-foreground flex items-center gap-2">
-                      <BrandIcon
-                        slug={providerMeta.providerBrandIconSlug}
-                        fallbackText={providerLabel}
-                        fallbackColor={providerMeta.providerBrandColor}
-                        size="sm"
-                      />
-                      <span>{providerLabel}</span>
-                      <span className="inline-block rounded-full px-2 py-0.5 text-xs font-medium bg-muted text-muted-foreground">
-                        {models.length}
-                      </span>
-                    </h3>
-                    <div className="grid gap-3 sm:grid-cols-2">
-                      {models.map((m) => (
-                        <div
-                          key={m.id}
-                          className="rounded-xl border border-border bg-card p-4 space-y-2"
-                        >
-                          <div className="flex items-start justify-between gap-3">
-                            <div className="flex items-center gap-2 min-w-0">
-                              <BrandIcon
-                                slug={m.brandIconSlug ?? m.providerBrandIconSlug}
-                                fallbackText={m.providerLabel ?? m.provider}
-                                fallbackColor={m.providerBrandColor}
-                                size="sm"
-                              />
-                              <span className="font-medium truncate">{m.name}</span>
-                              {m.contextWindowDisplay && (
-                                <span className="font-mono text-[10px] text-muted-foreground shrink-0">
-                                  {m.contextWindowDisplay}
-                                </span>
-                              )}
-                            </div>
-                            <Badge variant="secondary" className={`${TIER_STYLES[m.tier]} shrink-0`}>
-                              {m.tier}
-                            </Badge>
-                          </div>
-                          <div className="font-mono text-xs text-muted-foreground truncate">{m.id}</div>
-                          {m.description && (
-                            <p className="text-xs text-muted-foreground">{m.description}</p>
-                          )}
-                          <div className="flex flex-wrap items-center gap-1">
-                            {m.bestFor && (
-                              <Badge variant="outline" className="text-[10px] font-normal">
-                                {m.bestFor}
-                              </Badge>
-                            )}
-                            {m.tags && m.tags.map((tag) => (
-                              <Badge key={tag} variant="outline" className="text-[10px]">{tag}</Badge>
-                            ))}
-                          </div>
-                          <div className="flex items-center justify-between gap-2 pt-1">
-                            {m.costRange ? (
-                              <span className="text-[10px] text-muted-foreground">{m.costRange}</span>
-                            ) : <span />}
-                            <Button
-                              variant="outline"
-                              size="xs"
-                              onClick={() => setAsDefault(m.id)}
-                            >
-                              Set as Default
-                            </Button>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )
-              })
-            )}
-        </div>
-      )}
-
-      {tab === 'aliases' && (
-        <div className="space-y-4">
-            <div className="flex items-center justify-between">
-              <p className="text-sm text-muted-foreground">
-                Model aliases from <code className="text-xs">agents.defaults.models</code>
-              </p>
-              <Button variant="outline" size="xs" onClick={prepopulateAliases}>
-                Pre-populate Defaults
-              </Button>
-            </div>
-
-            <div className="overflow-hidden rounded-xl border border-border">
-              <Table>
-                <TableHeader>
-                  <TableRow className="bg-card">
-                    <TableHead>Alias</TableHead>
-                    <TableHead>Target Model</TableHead>
-                    <TableHead className="w-[80px]" />
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {Object.entries(aliases).length === 0 ? (
-                    <TableRow>
-                      <TableCell colSpan={3}>
-                        <InlineEmpty message="No aliases defined" />
-                      </TableCell>
-                    </TableRow>
-                  ) : (
-                    Object.entries(aliases).map(([name, target]) => (
-                      <TableRow key={name}>
-                        <TableCell>
-                          <code className="text-sm font-medium">{name}</code>
-                        </TableCell>
-                        <TableCell>
-                          <span className="text-sm text-muted-foreground font-mono">{target}</span>
-                        </TableCell>
-                        <TableCell>
-                          <Button
-                            variant="destructive"
-                            size="xs"
-                            onClick={() => deleteAlias(name)}
-                          >
-                            Delete
-                          </Button>
-                        </TableCell>
-                      </TableRow>
-                    ))
-                  )}
-                </TableBody>
-              </Table>
-            </div>
-
-            {/* Add alias form */}
-            <div className="flex items-end gap-3 rounded-xl border border-border bg-card p-4">
-              <div className="flex-1 space-y-1">
-                <label className="text-xs font-medium text-muted-foreground">Alias Name</label>
-                <Input
-                  value={newAliasName}
-                  onChange={(e) => setNewAliasName(e.target.value)}
-                  placeholder="e.g. opus"
-                />
-              </div>
-              <div className="flex-1 space-y-1">
-                <label className="text-xs font-medium text-muted-foreground">Target Model</label>
-                <ModelSelect
-                  value={newAliasTarget}
-                  onChange={setNewAliasTarget}
-                  models={modelOptions}
-                />
-              </div>
-              <Button onClick={addAlias} disabled={!newAliasName.trim() || !newAliasTarget.trim() || !modelsReady}>
-                Add
-              </Button>
-            </div>
-        </div>
-      )}
-
+      {tab === 'agents' && <AgentsTab m={m} />}
+      {tab === 'available' && <AvailableModelsTab m={m} />}
+      {tab === 'aliases' && <AliasesTab m={m} />}
       {tab === 'routing' && <RoutingTab m={m} />}
-
       {tab === 'spend' && <SpendTab m={m} />}
-
     </div>
   )
 }
